refactor(organization): restructure organization dashboard
